refactor: 코드리뷰 반영하여 패키지 구조 변경
