+++ product/src/main/java/com/core/exception/GlobalExceptionHandler.java
@@ -1,4 +1,4 @@
-package com.emotionalcart.product.presentation.exception;
+package com.core.exception;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
