--- product/src/main/java/com/emotionalcart/product/application/dto/request/GetProductReviewsRequest.java
@@ -1,9 +1,0 @@
-package com.emotionalcart.product.application.dto.request;
-
-import lombok.Data;
-
-@Data
-public class GetProductReviewsRequest {
-    int pageSize = 10;
-    int pageNumber = 0;
-}
